# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Employee Attendance System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a new employee (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "409":
          description: Employee already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /qr/generate:
    post:
      operationId: generateQrToken
      tags: [qr]
      summary: Generate a new QR token (admin only, expires in 5 min)
      security:
        - bearerAuth: []
      responses:
        "201":
          description: QR token generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrToken"
  /qr/current:
    get:
      operationId: getCurrentQrToken
      tags: [qr]
      summary: Get the current active QR token (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active QR token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrToken"
        "404":
          description: No active token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Employee:
      type: object
      required: [id, employeeId, name, email, role, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, employee]
        createdAt:
          type: string
          format: date-time
    EmployeeInput:
      type: object
      required: [employeeId, name, email, password, role]
      properties:
        employeeId:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [admin, employee]
    EmployeeUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, employee]
        password:
          type: string
    QrToken:
      type: object
      required: [id, token, expiresAt, createdAt]
      properties:
        id:
          type: integer
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...