# lib/api-zod/src/generated/types/employeeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Employee Attendance System API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeInputRole } from './employeeInputRole';

export interface EmployeeInput {
  employeeId: string;
  name: string;
  email: string;
  password: string;
  role: EmployeeInputRole;
}