# lib/api-zod/src/generated/types/qrToken.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Employee Attendance System API
 * OpenAPI spec version: 0.1.0
 */

export interface QrToken {
  id: number;
  token: string;
  expiresAt: Date;
  createdAt: Date;
}